finalize some things. fix license. remove very slow test case ( case always passed, it was just slow ).
