+++ .gitignore
@@ -7,7 +7,8 @@ __pycache__/
 *.so
 
 # Lock files
-.lock
+*.lock
+package-lock.json
 
 wheelhouse/
 # Distribution / packaging
@@ -42,8 +43,7 @@ wheelhouse/
 pip-log.txt
 pip-delete-this-directory.txt
 node_modules/
-package.json
-package-lock.json
+
 # Unit test / coverage reports
 htmlcov/
 .tox/
