+++ .pre-commit-config.yaml
@@ -8,9 +8,3 @@ repos:
         pass_filenames: false
         always_run: true
         types: [python]
-    -   id: ruff
-        name: ruff
-        entry: ruff check --fix
-        language: system
-        types: [python]
-        pass_filenames: true
