+++ faster_outlines/fsm/__init__.py
@@ -1,9 +1,10 @@
 from .regex import create_fsm_index_end_to_end
 
-from faster_outlines.fsm.fsm_utils import TokenVocabulary, Write, Generate
+from faster_outlines.lib import TokenVocabulary, Write, Generate
 
 
-from .guide import RegexGuide, LazyVLLMRegexGuide
+from .guide import RegexGuide
+from .vllm_guide import LazyVLLMRegexGuide
 
 __all__ = [
     "TokenVocabulary",
