--- rust/fsm_utils/src/lib.rs
@@ -1,29 +1,0 @@
-// Copyright 2024 Nathan Hoos
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#[macro_use]
-extern crate anyhow;
-
-mod tokenizer_index;
-mod environment;
-pub mod lazy_index;
-mod caching;
-pub mod types;
-pub mod vocab;
-mod bindings;
-mod atomic_wait;
-mod sp_decode;
-
-#[cfg(feature = "python_bindings")]
-pub use crate::bindings::fsm_utils;
